+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2026-06-23
+
+### Added
+
+- Added visible "Sponsored by ReadyPixl" placements to the global footer and
+  project detail sidebar, using the official ReadyPixl wordmark and a sponsored
+  outbound link.
+
 ## [0.4.0] - 2026-06-22
 
 ### Added
@@ -176,7 +184,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - N/A
 
-[Unreleased]: https://github.com/iScale-Builders/iscale-builders/compare/v0.1.7...HEAD
+[Unreleased]: https://github.com/iScale-Builders/iscale-builders/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/iScale-Builders/iscale-builders/releases/tag/v0.4.1
+[0.4.0]: https://github.com/iScale-Builders/iscale-builders/releases/tag/v0.4.0
 [0.1.7]: https://github.com/iScale-Builders/iscale-builders/releases/tag/v0.1.7
 [0.1.6]: https://github.com/iScale-Builders/iscale-builders/releases/tag/v0.1.6
 [0.1.5]: https://github.com/iScale-Builders/iscale-builders/releases/tag/v0.1.5